Performance optimizations in motion statistics; see full commit message
This commit also contains performance testing code and correcting testing
code, both to be removed in the next commit. It was tested on a post-
round 2 Australs database, so it retrieved statistics for six motions.
The results of the new code was compared to that of the old code, and
the numbers are identical. The new code was clearly faster than the old
code. To test time taken, each function is called three times in the
code, and these numbers indicate the total time taken in seconds for
all three calls:

new 2.27844405174
old 4.11044096947

new 2.27196907997
old 3.52677798271

new 1.95923399925
old 4.01483678818

new 2.27844405174
old 4.11044096947