Fix availability bug with checkboxes
The jQuery filter for checkboxes would also pick up on extraneous checkboxes,
like the ones in the Django debug toolbar sidebar. This would then cause an
error saving availabilities, because the extraneous checkboxes would have
a data-target of "NaN". This commit adds a class to checkboxes so that only
relevant ones are checked by the relevant JavaScript code.