Remove speaker from SpeakerScore unique_together; fix associated bugs
The bugs would cause duplicate speaker scores to be saved when you
change the speaker order in an existing ballot, because it would
create a new SpeakerScore without deleting the old one, since
because the speaker has changed it would not recognize it as the
same object. This wouldn't have caused anything to crash, because
we didn't previously have any get() calls, but it would presumably
have made some calculations give incorrect results.

Furthermore, I'm pretty sure speaker scores are meant to be unique
per debate-team-position-ballot -- you can't have two scores for
the same position in the same debate, even if speakers are different.