Remove raw SQL method of annotating availability
This method is better at encapsulation, in that it doesn't require
other apps to declare
    round_availabilities = GenericRelation('availability.RoundAvailability')

It also has the advantage of being able to collapse availability
retrieval and instance retrieval into a single query: Django, much to
my chagrin, doesn't provide an easy way to annotate with an EXISTS clause.

However, it comes at a cost: it violates DRY, the raw SQL clause itself is
messy because we have to retrieve ContentType.get_for_model(self.model)
ourselves, and it's prone to breaking with subtle changes in the models.
The nail in the coffin in this case is that Adjudicator doesn't have
an 'id' field: it relies on 'person_ptr_id'. Then either the raw SQL
generation needs to be told this, or it needs to infer it, otherwise it
will find that 'participants_adjudicator'.'id' doesn't exist. That's just
a bit too fragile for us to be able to rely on a RawSQL clause, especially
in an app that's trying to take advantage of the power provided by
the contenttypes application.