Make most CharFields non-nullable (see full commit message)
CharFields aren't meant to have null=True set, as explained here:
https://docs.djangoproject.com/en/1.10/ref/models/fields/#null

However, when we want to avoid enforcing a uniqueness constraint
on a CharField, we can use a nullable CharField to achieve this:
https://code.djangoproject.com/ticket/4136

So where there is a unique CharField we use null=True, but if
there's no uniqueness constraint we should follow the Django
convention of storing as empty strings, so that there's only
one representation of having no data in the field.

Also, this commit does #382 for adjfeedback.