Fix race condition in prefetch_ballotsets (see message)
The race condition occurs because, if ballotsubs is a QuerySet,
a query filtered on "ballot_submission__in=ballotsubs" or
"debate__ballotsubmission__in=ballotsubs" would pass the QuerySet
to the filter, and the query would then filter using a nested
SELECT statement, not using a list of BallotSubmission IDs. This
means that the submissions actually in ballotsubs may change on
each of the queries in prefetch_ballotsets, causing a race
condition, as other code assumes that it will always be able
to find the relevant object retrieved from earlier queries.

The fix is to force evaluation of the ballotsubs queryset into
a list, and then pass the list into the filters, so that it
filters by IDs, and the IDs will be fixed at the beginning of
the function rather than changing every time.