Set migration order of conflicting migrations
Note: The two migrations numbered draw/0020 actually conflict, so
can't be merged. I resolved this by specifying that the migration
from develop should be done before the migation on the feature/bp
branch. This will mess up any existing installations on the
feature/bp branch, because their migration history will now be
inconsistent with the (new) migration tree.

The order has to be resolved in this way for two reasons: (1) it
makes more sense logically (the develop migration changes the
convention from 'A'/'N' to 'aff'/'neg'; the feature/bp migration
then adds the BP positions), and (2) installations on the develop
branch that need to be migrated when BP support is rolled in will
obviously have the develop one, but not the feature/bp one, so
it has to be this order to avoid an inconsistent migration history.

Finally, note that owing to an oversight, the migration on the
develop branch and now earlier in the tree (0020_auto_20170713_1530)
was actually created after its feature/bp counterpart that is now
later in the tree (0020_auto_20170618_1549).