Abolish DRAW_FIRSTBREAK in favor of just detecting it directly
That is, in situations where it matters whether or not it's the first
break round, it now just checks whether the previous round was also
a break round.

Also, rename DRAW_BREAK ('B') to DRAW_ELIMINATION ('E'), consistent
with our terminology elsewhere in the source code.

It still enforces that elimination rounds must have draw type
DRAW_ELIMINATION.