Add ability to customize ballot/feedback emails
The private URLs to the ballots and feedback are sent to their
respective participants through email through a template, which is then
mail-merged with the information. This change makes these templates
customizable, allowing for further instructions or a different message
to be used.

A few variables are used for this mail-merge, with the same names as
previous. The tournament name is inserted directly into the message,
without the need for the merge.

As a consequence, there is a preview for the email messages before they
are sent.