Send receipt email to judges with their ballot
This change sends the judges of the debate a confirmation email that
their ballot(s) for the current round have been received (with the
respective judge's scores). The email is sent after the form submission.
The message of the email are tournament settings under 'Data Entry'

This is very useful to give the judges a record of their votes, and also
to give them an opportunity to re-check and validate their scores if
there is a problem.

The message is somewhat customizable, with keywords shown in the prefs.
However, the format of the variables these keywords represent are only
localizable.