Add field to render judge scores ineffective
This change adds a field `ignored` to the judge feedback application.
What this new field does is excludes the score from being taken into
the score calculations, without it being entirely discarded.

Forms can be marked as ignored in the admin/assistant submission form,
but not through public ones. Ignored ones are marked as such in the
admin views.