Add options for email sending
This commit adds a new configuration section to control for what should
email be sent out. New boolean settings were added to this section, as
was the message content for ballot receipts. The main option also
dictates whether to show the links towards sending private URL emails to
teams/adjudicators.

An error flow for the ballot receipts was added to add failure/success
messages to indicate the result of the email routine in the admin/assist
views with the confirmation of ballot sets. An indication that emails
will be sent is also added to the submit button.

A missing gettext was added in the preferences.