Allow for team points to be emailed to teams
This commit adds a method to email the current team points to the
members of each team. The format used is similar to other emailing
methods with a customizable message. Calling this method, a new button
has been added before starting a new round allowing for these messages
to be sent to participants of the last round.

Due to the difference between how points are given in BP and 2-team
formats, there is a different pre-set message for BP, as does the
button.