Begin Notifications application
This adds a subclass of EmailMessage in a new "Notifications" app. This
new subclass allows for messages to be better customized for tournaments
with Reply-to, custom from names, and integrated logging though a method
as_model(). New settings were added for the Reply-to fields, with a new
table that supersedes the private URL records. This model has been put
to use where applicable.