Give all participants a private personal URL
All participants in a given tournament will have their own private URL
key for use within the tournament. This is an important for the
security of these keys and also for convenience. With this in mind, the
Person, Adjudicator, and Team models were modified to have the URL key
as a field of Person, and removing it from Teams. This allows for a more
granular control over who in each team is submitting feedback. Team
feedback is kept as being a team submission through the Speaker->Team
relation.

As there is getting to be more and more uses for these private keys, a
splash-page for these individuals is created within the PrivateURLs
application which contains links towards ballot/feedback submission
forms. Printed pages and emails will only have that new link, reducing
the amount of printing/emailing.