Expand the sent messages record
This commit renames the MessageSentRecord model to SentMessageRecord
while adding new fields for the raw email address, the context in the
message, and the message itself. With that, the TournamentEmail module
has been expanded to include the mail-merge features itself.

Extra options for email type and method were added in anticipation of
new features.

The Postgres JSON field is used to contain the context in a JSON format.