Fix caching problem in personalized landing pages
The underlying issue was that the page was being cached (as I had
thought). Saw 304s being sent out as the landing page. To fix, I created
a new mixin which is like the PublicTournamentPage mixin but without the
caching. Caching can't be used well for personalized pages, which
PublicTournamentPageMixin wasn't for (now a subclass of personalizable).

Also fixed a few typos relating to the feature.