Add average individual score to Team metrics
The average individual score (AIS) is like the average speaker score
(ASS), but it returns values within the tournament's scoring range as it
averages the speaker scores between all speakers in all rounds, and not
just between all rounds. It does not include reply speeches however as
they may use a different scale.