Minor touch ups to the notification style
The secondary style is intended essentially for passive information — i.e. the presence of motions. The default/blue style is more for 'pertinent information' to a particular action (in this case the actions that releasing will trigger).
I've also expanded the email_text, both to indicate that this is a preference that can be disabled. 'Matchup' to my mind is also slightly US-centric jargon that would be unfamiliar elsewhere, so I've more explicitly spelled out what is happening.