Rename "average speaker score"
This commit renames the average speaker score to the average total
speaker score, which was how it was described. It replaces the
unfortunate acronym to "ATSS". However, its internal name remains
unchanged to keep compatibility.

This is needed with the introduction of the average individual score,
where its most intuitive name conflicted with this metric, and changing
that one to average speaking score would be confusing.