Create new option to enable motion re-use
This commit adds a new option for tournaments that wish to re-use
motions in a new section. This section also contains other motion-type
settings like 'enable-motions'.

By enabling the setting, a new form appears in the edit motions view to
select available motions, and a button to easily re-use the motions of
the previous round appears.