Change ballots graph to area chart
Reverts to the form of an earlier iteration of the graph, including
colours, but with d3.v4. An area chart is more suited for this kind of
data which is based on time, not discrete. With the changes, the
generation should also be simplified without the need for the "total"
key.

This also partially reverts 0555bcc68b39df57bd7f3be905568f33a793ffb2 to
have the "square" effect on the chart.