Fix npm cp-i18n and cp-fonts commands
Both commands would previously only work if the static/ directory
was clean, i.e., if the target directories didn't exist. Otherwise,
they would copy the entire source directory, to a new directory
within the target directory, that is, to:
    static/jsi18n/jsi18n/
and
    static/fonts/Inter UI (web)/
respectively, when the intention was to copy the *contents* of the
source directories to static/jsi18n and static/fonts.

This change modifies the relevant commands so that they will work
whether or not the target directory already exists, by using
`mkdir -p` before the `cp` command.