Make private URL email message in preferences
This removes the need for the separate page, with a warning list for
the participants where email failed. Now the message is customizable
though the email preferences, and the link to the email form is now a
form button which deactivates if everyone reachable has been sent their
key.

The delta check for if the key has changed does not work and so has been
removed.