Remove get_cached_institution
In my tests, hitting the cache for each institution was responsible
about 17 seconds, out of 32 seconds for the page, i.e., a majority
of the load time (according to the debug toolbar profiler). The
better strategy is to use select_related() to select the institution
and region of teams in the original SQL query, which is what
DrawForDragAndDropMixin.annotate_draw() already did
(see tournaments/mixins.py:395).

Since get_cached_institution() doesn't seem to be used anywhere else,
I removed it entirely. However, there might be other places that
hit the region of teams often; if so, those queries should have
select_related('institution__region') added to them.