Performance improvement in serialization methods
This optionally allows Debate.serialize() to take a `tournament`
argument, which, if provided, it uses for all of its tournament
preference lookups. This means it can use the instance-level cache
rather than having to look up from the actual cache for each
debate (because debate.round.tournament is a different Python
object for each debate).