Add QR codes to private URL sheets
It is awkward to type in a coded URL. To make it easier, QR codes allow
for a URL to be encoded in a 2D barcode, redirecting to a URL when
scanned with a cell-phone. The QR codes are generated in the client with
qrcode, a new python dependency.

As this is done server-side, it will load slower, but PDFs are generated
quicker.