Add team data to ballot receipt
This adds the team's ranking and total score to the ballot receipts for
easier comparison. In addition, now the codes for sides are used to save
on space.

Team names vs codes have been corrected by checking the preferences and
using what is shown in the public draw page. The as_dicts() function is
no longer used with the email generator going directly to the scoresheet
and DebateResult objects.

In consensus ballots, all the members of the jury will be sent the
ballot receipt.