Use channels to send notifications
Email generators will now only create tuples of (context, user) to be
made into emails to be sent off. In this new setup, the action of
requesting emails is sent to an asynchronous channel which then calls a
generator. This should reduce the number of calls to the database, and
make emails non-blocking.

The side-effect is that success/failure messages cannot be sent as
easily, and has not been attempted here. Rather, messages notifying
that messages *are queued* have been made instead.