Add parameter to email generators for people list
In order to give tab directors more granular power over who to message,
an array argument is added to the generators with the IDs of the people
to be sent messages.

For now, this is just the default, but a new view will allow for the
selection of a different (but possible) group.