Add views to send templated emails
This commit adds two views (one for round-related and the other for
tournament-related emails) that allows for the recipients to be selected
from the list of relevant parties. The relevant parties are selected by
default.

The email message is integrated within the dynamic_preferences framework
and auto-updates. These settings have been removed from the main options
interface to reduce the load there.