Add registry for and revive old adjudicator allocators
Deleted the STab allocator; it seemed beyond repair. In theory,
the new registry would allow for future provision to choose which
allocator to use, a bit like break generators now.

Added a management command `allocateadjudicators`, which runs
the allocator and can take some command-line options, in particular
to do a dry run and to select which allocator to use.

Extended AdjudicatorAllocation to be able to handle both debates
and preformed panels (using the latter is a work in progress).