Further subclass email views
This commit separates templated email views into the specific events.
The new subclasses are placed in the application which collects the data
to be disseminated. The consumer now takes the SentMessageRecord event
as the message parameter.

The consumer also has two new parameters for the subject/body content.

This also reduces the number of switches between types with the 2nd
"email choices". Also, missing emails are highlighted.