Create new models for notifications
Two new tables are created for the notifications app in preparation for
integration with SendGrid's event webhook. These new tables will reduce
repeated information while logging messages and now events.

- BulkNotification
  This table contains information relating to the group of messages that
  were sent at once (as is the norm for bulk messages). The fields are
  mostly taken from the SentMessageRecord table, with the migration
  transposing the affected fields (grouping by timestamp)
- EmailStatus
  When SendGrid sends a POST, the event will fall to this table for use
  within the application. Emails are identified by their Message-ID
  which has been added to SentMessageRecord (with a migration step to
  fetch available Message-IDs).

Message-IDs could have preferably been used as a primary key to
SentMessageRecord, but some records might only have the body of the
message and not the headers (which has the Message-ID).

This commit also converts the old uses of SentMessageRecord to the new
uses.