Pass tournament in AISS metric annotation
The "Average Individual Speaker Score" ("AISS") metric requires the
'substantive_speakers' preference in order to exclude reply scores (as
they may have a different range). Or, the tournament object is not
provided so this commit takes the tournament object from the first
object of the applicable queryset (Team) for use in the metric.

Ghost scores are now ignored as well.