Make recipient nullable in sentmessagerecord
As there is still an email address attached, the recipient field is not
vital. Also reversed the ordering and added the participants' names as
an ordering field and added get_latest_by for EmailStatus.

Even though Message-IDs are valid emails, they are not recognized by
Django, so the field has been changed to a simple CharField.