Create emails as HTML
This commit makes the email templates and custom emails HTML-based with
the Summernote widget. This allows for further customization.

To keep compatibility, a new function was created to strip all the tags
in the message, giving a plain-text version as well.

For consistency, the message attribute of the custom email event has
been renamed to body.