Set REGISTRY_MODULE setting of dynamic preferences
Turns out the file rename in 251509f16 wasn't just an arbitrary
name; django-dynamic-preferences makes an assumption about where
it will find all of the preference definitions for autodiscovery.
I don't really like the long file name, so this commit sets
a django-dynamic-preferences setting to change where it looks for
the preferences instead.