Split the "In this Round" card
This commit splits the components of the "In this Round" card from team
and adjudicator pages into separate files to reduce code duplication. In
addition, a new parameter for the grammatical gender was added between
the third and second person. Fixed some i18n omissions too.

The box is now on private URL landing pages. It uses the second person
(motivation for the new parameter). With the information showing when
the draw is released, the page is useful (active) even if no actions are
available.