Add debate table to private URL landing
Now the private URL landing pages has a table for the debates in which
the participant was involved. For adjudicators, this includes a link to
their ballot, which would normally be private (but was sent as an email)

The action links have also been moved up to the top of the page for
easier access.