Remove atomic-type of notifications migration
The migration comported both a schema and data migration, which caused
an error when filling in message-ids. This commit corrects the behaviour
by specifying the migration as non-atomic.

The commit also makes the BulkNotification data migration check for
successive events, and not timestamps. It also removes the Regex to look
for Message-IDs in favour of Python's email module.