Implement team scores by adj table
This commit adds a new table to the Results application to hold team
scores by individual adjudicators (TeamScoreByAdj) as a counterpart to
SpeakerScoreByAdj and TeamScore. Its use is primarily for by-adj
elimination ballots. The uses are not yet implemented.

It has these fields:
 - Ballot submission
 - Debate-Adjudicator
 - Debate-Team
 - Win?
 - Margin
 - Score

The points field was not included as only two-team formats can use this
table and the win column can serve the same role. The margin field was
added as it could signal greater dissent from an adjudicator than
another.

A database migration has also been added to both create this model and
to populate it with existing SpeakerScoreByAdj records. All of the
current data in this table will be redundant with existing data.