Bump python to 3.6.8
This is requried to install the latest version of channels, so it seems like we need to do so at some point. This also aligns us with the heroku supported runtime.

Windows installation instructions have changed and require testing, but theoretically should be fine. Forcing an install of the 32-bit version is not ideal; but I don't think worth worrying about.