Clarify i18n/l10n in contributing reference
This commit adds detail on how to work on i18n and l10n in the
contributing page of the documentation. Detail is added specifically
about the differences in handling JS and Django files.

Unsure of whether only the master english `djangojs.po` file or all
language files must be updated with new strings when added (but suppose
all of them must as the manual step of `makemessages`)