Implement pull-up filter by draw strength
This commit implements a filter for pull-up eligibility as having the
lowest draw strength of one's bracket. This allows to compensate for
if a team had easier match-ups in previous rounds (even if power-paired)
and so had an advantage.

By draw strength in this context, the average of team scores (by speaker
score) from previous rounds is used. The average is used instead of the
sum to prevent teams who had been absent a round from always being
pulled up. Scores are used instead of wins as brackets are already based
on wins.