Rephrase assertResponseOK
Currently, because it catches exceptions, it also catches the
AssertionError associated with a standard test failure, before
reraising a new exception using self.fail(). This modifies it
so that it just checks for equality directly and calls self.fail(),
or if an unexpected exception occurs, it will just be handled as
usual by the test framework.