Rewrite draw strength annotator to query for opponents directly
Previously, this relied on the populate_opponents() method in
draw/prefetch.py, which is designed to produce a single opponent
per DebateTeam given to it (mainly for use in two-team formats).
This doesn't work in BP (see #1071). Furthermore, draw strength
doesn't require information on individual debate teams, it just
requires a list of all opponents seen by a team, which can be
queried more directly.

Fixes #1071