Permit empty standings precedence and warn user accordingly (#1149)
* Make standings generation more robust #1108

This rewrites metricgetter() to avoid using Python's itemgetter(),
so that the function it returns will always return a tuple. This
simplifies a lot of handling, because it avoids the special case
where itemgetter(item) called with one argument returns a single
item rather than a tuple of length 1.

As a consequence, the case where no metrics are specified now
shouldn't crash. Also added standings test for this case.

* Fix bug where extra metrics cause standings to crash

Also added a unit test for this scenario

* Add user alerts for blank standings precedence #1108