Notifications: Deal when recipient is null
Recipients to a SentMessage can be null, which may cause a key error
when trying to access their attributes. This happens in the email
status view. A check is added for this case.

This also affected the repr of emails, which now uses the email when
the recipient is None, falling back on the Message-ID.