Travis: Bump PostgreSQL version to 11
This commit is a squash of about 15 commits, all with various
efforts to figure out how to bump PostgreSQL to version 11 on
Travis.

Some people include lines to uninstall the pre-installed versions,
i.e. versions 9.x, but that's not really necessary. But it is
necessary to copy the old pg_hba.conf file over, otherwise peer
authentication for the user 'postgres' fails, presumably because
it's not committed.