Restructure DebateResult/Scoresheet
This commit reduces the number of various DebateResult classes by making
them agnostic to the number of teams. With that, the base class branches
to "By Adjudicator" and "Consensus," with a mixin included for their
subclasses to support scoring.

For this, scoring is assigning speaker scores, which are then combined
to create team scores. Teams may not be scored without speakers.

This greatly simplifies the inheritance structure of DebateResult.
The scoresheet selection for each type of DebateResult is in the
DebateResult class. To support BP in this structure, the `winners()`
attribute of scoresheets is now a list of winning sides. There is an
assertion to prevent an incorrect number of winners though.

Also, a specific scoresheet for BP Elimination has been created.

No new options to access DebateResults without scores has been created
in this patch.

Tests are also updated to take the new structure, but no new tests have
been added as of this commit.

Finding the proper DebateResult class has been spun-off from the factory
function.