Update uses of DebateResult
As a consequence of the previous commit, uses of DebateResult in certain
places needed to be updated. This commit updates its use for ballot
notifications (using isinstance) and the dropping of "advancing" in
favour of the generalization "winner(s)" in results.

This commit does not update ballot forms, which will be dealt with
separately.