Make importtournament command use UTF-8 by default
This is a backport of d78b9b98eb3621dfb560ca567d4e3bdf083714ff from
develop to release/2.3.0.

Importing the demo tournaments on Windows fails, because the preferred
encoding on this OS (which might also be locale-dependent) seems to be
CP1252. However, all of the demo datasets are encoded in UTF-8.

This commit:
 - adds an option to the importtournament management command allowing
   the user to specify an encoding
 - defaults to UTF-8 if an encoding isn't specified
 - in the UI demo dataset import view, hard-codes the encoding to
   UTF-8

This also adds a check that the demo dataset source requested in the
POST is one of the permitted ones.

Since the import of user CSV files is currently only possible through
the command line, an update to documentation isn't necessary. However,
when #422 is implemented, there should be a note indicating that UTF-8
encoding is expected (or an ability for the user to choose, or some
sort of auto-detection).

Thanks to Manu García for the report.

(cherry picked from commit d78b9b98eb3621dfb560ca567d4e3bdf083714ff)