Results: Refactor forms
This commit refactors the result forms to be orthogonal to the Debate-
Result classes. In that, mixins are used to distinguish forms using (or
not) scores.

The DebateResult class has also been added as a parameter for the forms
with a correspondence dict between forms and results in views.py.

A new preference allows for speaker scores to be omitted from ballots
depending on the round stage.