i18n: Use humanize for cardinals
This commit uses the django.contribs humanize module to get the ordinal
form of numbers when used in apps.

Also, numeral forms of numbers have been removed from localization as
number translation does not seem prevalent and is not supported by the
rest of Django.

The ballot vue templates have been updated for this, taking ordinals
from Django. Used this opportunity to make loops for the numerals.