Show missed speeches message if limit is 0
The filter for the number of speeches a speaker can miss and still be in
the tab has a default of -1 to include all speakers. The choice of 0
(can't miss any debate) is a valid option, but was unaccounted for in
the messages about the option.

This changes the comparisons to greater than *or equal to* 0.