Rename "test score" to "base score"
As mentioned in https://github.com/TabbycatDebate/tabbycat/pull/762,
"test score" could be a misnomer for tournaments that do not use
adjudication tests to determine first score. For that, the term "base
score" is more general for those cases, yet would still be
understandable for tournaments with testing.

Migrations of models and methods using the old name have also been made
for consistency of the term.