Associate venue categories with tournaments
In addition to venues themselves, their categories should be associated
with particular tournaments. This prevents the formset from being over-
loaded with other tournaments' categories, which also has the side-
effect of removing other tournaments' venues from their categories with
saves.

In addition, the help text for venues and their tournament is removed
as shared venues/adjudicators were removed with divisions, as well as
setting tournament as blank=False.