Create Break/Speaker Category endpoints
These API endpoints are as a proof-of-concept with the API application.

The endpoints include endpoints to list, delete, and add speaker and
break categories to a tournament, including modifying the team/speaker
sets. These endpoints use the slug field of the models and not their
IDs, although they use speaker IDs.

Using ViewSets for list and detail views, another PUT/PATCH view for
specific categories was also added to be able to set eligibility or add
participants to a category using PUT / PATCH respectively.