Create view to create/edit break categories
This commit adds a new interface to create and edit break categories,
modeled on the creation of speaker categories. While traditionally done
in the database area, the creation of break categories has more steps
than just creating a model object, so is best done interactively.

When creating a break category, rounds are now automatically created for
the number of teams breaking, using the same method for the open break
rounds as in tournament creation. In order to have the appropriate round
names, the round names have the category name, and the abbreviation has
the first letter of it.

A method for modifying existing objects has been also subclass-ed for
future use (if need to delete/append new rounds due to a change in
break size).

The documentation has also been updated to reflect this new procedure.