Replace "open-source" by "source-available" (#1338)
* Replace "open-source" by "source-available"

This is a distinction between the current licensing scheme being non-
libre, but the source code and development being done openly. This
change reflects the README which states it is not open-source (yet).

* Replace 'open-source' by 'source-available' in docs

This commit replaces 'open-source' by 'source-available' in the
Tournament Logistics documentation page, as the documentation is
probably with the same license as the rest of the tournament.