Mark private URL explainer as a warning
The private URL landing page contains an explainer of the page with a
warning about unauthorized access. Other pages, such as the private URL
view of scoresheets, have similar warnings. This commit standardizes the
type of the messages, changing 'info' to 'warning' on landing pages in
line with scoresheets.

This commit also removes the note about tournament rules in the
scoresheets as it is redundant with the explication why sharing the
credential is bad, without resorting to authority.

Also removed an unused if statement.