Rename "Venue" to "Room" in UI
This commit changes most instances of "venue" found in the user UI and
shown messages to "room," including for categories and constraints.

The term "room" seems more appropriate than "venue" as the latter
normally refers to a building, and not the specific rooms inside it,
although large rooms such as auditoriums (like for break rounds) can be
called venues in and of themselves. However, they are normally treated
as rooms with numbers and such.

As a start, this commit only changes the interface. To go deeper with
this change, we would need to rename models and fields, then the whole
"venues" application.