Broadcast when ballots are confirmed
When ballots get confirmed, they were not being broadcasted to the
consumer, and thus didn't update in the graph as identity was checked,
not the string. Also placed it in the previous (identical) if block and
set the comparison with the model (static) attribute and not per-object.

Also set the BallotStatusConsumer to send the status of the debate and
not from the cleaned data and reduced a DB call by sending the round_id
from the debate.