Append indicator of submitted feedback on adjudicator
This commit adds an indicator to the beginning of the adjudicator string
in the dropdown to show which adjudicators have already been given
feedback, as to prevent unexpected feedback.

The new query checks for the existence of confirmed feedback for the
round from the source, which is a DebateAdjudicator attribute, but then
assigned to Adjudicator objects as those are used in the strings.