API: Add extra steps to object creation
Instead of directly creating objects within the create() methods of
model serializers, this commit changes to use the super-class method, or
to defer to the proper serializer (for speakers in teams).

These methods are now also used to implement creation procedures, such
as adding the home institution conflict, setting code name/emoji if not
set, and adding general break categories.

In addition, more fields have been added to adjudicators.