Re-implement debate postponement
In brief, debate postponement is to signal that a ballot should get re-
checked or is not vital to the creation of the next round, and so you
can advance to the next round with postponed debates.

Debates can be postponed with or without a ballot; to postpone without
results, each debate will have a button on the results page in order to
postpone the debate. They can only be un-postponed by adding results,
and in the assistant view, it is automatic. The button will work if
there are results, as the status could be used to indicate that the
ballot should be re-checked.

Postponing a debate activates the ballot status consumer, which updates
the table.

A similar feature was previously available with divisions, but was
recently removed.