Move Backups doc page to Use section
"Tabbycat doesn't provide an in-built backup system; instead you should
create copies of your database directly." Thus backups isn't really a
feature of Tabbycat, and is oddly categorized under "Features" in the
documentation. Not denying the value of the page, the page should be
kept.

This change moves the page to the use menu, as I see these instructions
being more in line with considerations on the use of Tabbycat with the
DB/data and the application itself.