Add selectors for adjudicator positions in draw emails
This commit adds three selectors in the Email Adjudicators' Draw view,
for chairs/panellists/trainees, which can be used to give different
instructions to these classes. To do so, type "type" in the checkboxes
has been generalized.

In addition, the get_success_url() has been simplified to just the
pattern with the get_redirect_url() of the Tournament/Round mixins doing
the rest.

# Conflicts:
#	tabbycat/notifications/views.py