Remove URL parameter from team draw emails
Caused a TypeError when trying to send as URL was not in the "extra"
fields. Removed the URL field entirely as it is less useful to teams
than for adjudicators and it wasn't listed as available in the docs to
begin with.

Also added some CHANGELOG lines for the hotfixes.

Fixes BACKEND-1SG