Create permission for public access to API endpoints
This commit creates a new permission based on IsAuthenticatedOrReadOnly
meant for tournament-related API endpoints to check a setting and give
public read-only access if allowed there. Full access is still granted
for administrators. This will be useful if an API user needs to collect
publicly available data (like for archival or Black-Rod).

New mixins were created for this new permission as well for the general
"Authenticated or read-only", as it could be useful for strictly public
objects. (Split for "Is Admin or read-only)

In addition, the mixins and this permission have been placed in new
files.