Create list page for inactive tournaments
If there is an inactive tournament, it will be shown on a new page based
on the site index, with a list of the tournaments.

This also undoes the change so that if there are inactive tournaments,
the site index will be shown with a link to the "Archive" page, even if
there is just one tournament.